requirements.txt
Flask==1.1.1
six==1.14.0
Flask_WTF==0.14.3
WTForms==2.2.1
Flask_SQLAlchemy==2.4.1
numpy==1.17.2
protobuf==3.11.3
google-cloud-translate==2.0.0
gunicorn==19.9.0
